Solved Backwash Issue
Issue cannot be solved with a way just a an extra line in the bottom, it could not pass the time and memory tests. The solution to this problem is what adding another WeightedQuickUnionFind datatructure, and this one doesn't connect to the bottom, use this one to check isFull(), use origin one to check percolate().